DATA ANALYTICS SESSION 1 ASSIGNMENT 1 SUBMISSION
# Project Title

One Paragraph of project description goes here

## Getting Started

These instructions will get you a copy of the project up and running on your local machine for development and testing purposes. See deployment for notes on how to deploy the project on a live system.

### Prerequisites

What things you need to install the software and how to install them

```
Give examples
```

### Installing

A step by step series of examples that tell you have to get a development env running

Say what the step will be

```
Give the example
```

And repeat

```
until finished
```

End with an example of getting some data out of the system or using it for a little demo

## Running the tests

Explain how to run the automated tests for this system

### Break down into end to end tests

Explain what these tests test and why

```
Give an example
```

### And coding style tests

Explain what these tests test and why

```
Give an example
```

## Deployment

Add additional notes about how to deploy this on a live system

## Built With

* [Dropwizard](http://www.dropwizard.io/1.0.2/docs/) - The web framework used
* [Maven](https://maven.apache.org/) - Dependency Management
* [ROME](https://rometools.github.io/rome/) - Used to generate RSS Feeds

## Contributing

Please read [CONTRIBUTING.md](https://gist.github.com/PurpleBooth/b24679402957c63ec426) for details on our code of conduct, and the process for submitting pull requests to us.

## Versioning

We use [SemVer](http://semver.org/) for versioning. For the versions available, see the [tags on this repository](https://github.com/your/project/tags). 

## Authors

* **Billie Thompson** - *Initial work* - [PurpleBooth](https://github.com/PurpleBooth)

See also the list of [contributors](https://github.com/your/project/contributors) who participated in this project.

## License

This project is licensed under the MIT License - see the [LICENSE.md](LICENSE.md) file for details

## Acknowledgments

* Hat tip to anyone who's code was used
* Inspiration
* etc



1. Say True or False for the below statements:

• Prescriptive Analytics used to predict the future outcomes?

       Ans: - False. Predictive Analytics is used to predict future outcomes.

• Base R packages installed automatically?

       Ans: False - we have install basic & R studio. Relavent packages are installed automatically.

2. What is Recycling of elements in a vector?

       Ans If your two vectors are not equal in length then the shorter one will be recycled to match the longer vector.

3. Give an example of recycling of elements.

        Ans
        a =c(10,20,30)
        b=(1,2,3,4,5,6,) b(6)
        a+b
       [1]11 22 33 14 25 36
       Virtually the shorter vector a(3) is recycling its values to match with "a[10, 20, 30,10,20,30]" virtually